Some actual progress

+++ src/pages/Home.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import fire from "../services/Fire";
 import Player from "../Player";
-
+import "../App.css";
 import * as $ from "jquery";
 
 import Button from "@material-ui/core/Button";
@@ -23,12 +23,10 @@ const Home = () => {
   const [is_playing, setIs_playing] = useState("Paused");
   const [progress_ms, setProgress_ms] = useState(0);
   const [token, setToken] = useState(null);
-  const [UID, setUID] = useState(null);
-  const [username, setUsername] = useState(null);
-  const [spotifyID, setSpotifyID] = useState(null);
-  const [image, setImage] = useState(null);
-  const [registered, setRegistered] = useState(false);
-  
+  const [user, setUser] = useState(null);
+
+  var username = fire.auth().currentUser.displayName;
+  var UID = fire.auth().currentUser.uid;
 
   const [item, setItem] = useState({
     album: { images: [{ url: "" }] },
@@ -37,41 +35,56 @@ const Home = () => {
     duration_ms: 0,
   });
 
-  const user = {
-    UID : "UID",
-    username : "username",
-    spotifyID : "spotifyID",
-    image: "image"
-};
+  var userRef = fire
+    .database()
+    .ref()
+    .child(`users/${fire.auth().currentUser.uid}`);
 
-  var userRef = fire.database().ref().child(`users/${fire.auth().currentUser.uid}`) 
-  
   useEffect(() => {
     // Set token
-    let _token = hash.access_token
+    let _token = hash.access_token;
 
-    if(!localStorage.getItem('token')){
-      if (_token) {
-        localStorage.setItem('token', _token);
-        setToken(_token)
-        getCurrentlyPlaying(_token);
+    userRef.on("value", function(snapshot) {
+      if (snapshot.exists()) {
+        var data = snapshot.val();
+        setUser({
+          UID: data.UID,
+          username: data.username,
+          spotifyID: data.spotifyID,
+          imageURL: data.imageURL
+        });
+        
+        if (!localStorage.getItem("token")) {
+          if (_token) {
+            localStorage.setItem("token", _token);
+            setToken(_token);
+            getCurrentlyPlaying(_token);
+          }
+        } else {
+          setToken(localStorage.getItem("token"));
+          getCurrentlyPlaying(localStorage.getItem("token"));
+        }
 
-        if(!registered){
-          getSpotifyUser(token)
-          setUID(fire.auth().currentUser.uid)
-          setUsername(fire.auth().currentUser.displayName)
-          userRef.set(user)
-          setRegistered(true)
+
+      } else {
+        
+        if (!localStorage.getItem("token")) {
+          if (_token) {
+            localStorage.setItem("token", _token);
+            setToken(_token);
+            getCurrentlyPlaying(_token);
+            getUserRegistered(_token);
+          }
+        } else {
+          setToken(localStorage.getItem("token"));
+          getCurrentlyPlaying(localStorage.getItem("token"));
+          getUserRegistered(_token);
         }
       }
-    }
-    else{
-      setToken(localStorage.getItem('token'))
-      getCurrentlyPlaying(localStorage.getItem('token'))
-    }
+    });
   }, []);
 
-  function getSpotifyUser(token) {
+  function getUserRegistered(token) {
     // Make a call using the token
     $.ajax({
       url: "https://api.spotify.com/v1/me",
@@ -83,8 +96,20 @@ const Home = () => {
         if (!data) {
           return;
         }
-        setSpotifyID(data.id)
-        setImage(data.images)
+        var tempUser = {
+          UID: UID,
+          username: username,
+          spotifyID: data.id,
+          imageURL: data.images[0].url,
+        };
+        setUser({
+          UID: UID,
+          username: username,
+          spotifyID: data.id,
+          imageURL: data.images[0].url,
+        });
+
+        userRef.set(tempUser);
       },
     });
   }
@@ -107,7 +132,7 @@ const Home = () => {
       },
     });
   }
-  
+
   function getMyPlaylists(token) {
     // Make a call using the token
     $.ajax({
@@ -118,14 +143,14 @@ const Home = () => {
       },
       success: (data) => {
         if (!data) {
-          alert('no data')
+          alert("no data");
           return;
         }
-        alert(data.items[0].name)
+        alert(data.items[0].name);
       },
     });
   }
-  
+
   function getUserPlaylists(token) {
     // Make a call using the token
     $.ajax({
@@ -136,66 +161,69 @@ const Home = () => {
       },
       success: (data) => {
         if (!data) {
-          alert('no data')
+          alert("no data");
           return;
         }
-        alert(data.items[0].owner.display_name)
+        alert(data.items[0].owner.display_name);
       },
     });
   }
-
+if(token) {
   return (
     <>
-      <h1>{fire.auth().currentUser.displayName}</h1>
+      <h1>{user.username}</h1>
+      <div className="profile__img"><img src={user.imageURL} /></div>
       <Button onClick={() => fire.auth().signOut()}>Sign out</Button>
-      <Button onClick={() =>getUserPlaylists(token)}>Playlists</Button>
-      {token && (
-                <Player
-                  item={item}
-                  is_playing={is_playing}
-                  progress_ms={progress_ms}
-                />)}
+      <Button onClick={() => getUserPlaylists(token)}>Playlists</Button>
+      <Button onClick={() => console.log(user)}>User</Button>
+      <Player item={item} is_playing={is_playing} progress_ms={progress_ms} />)
     </>
   );
+}
+else {
+  return (
+    <div>Loading...</div>
+  )
+}
+  
 };
 
 export default Home;
 
-
 // function getRefreshToken(code) {
-  //   // Make a call using the token
-  //   $.ajax({
-  //     url: `https://accounts.spotify.com/api/token?grant_type=authorization_code&code=${code}&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2Fregister`,
-  //     type: "POST",
-  //     beforeSend: (xhr) => {
-  //       xhr.setRequestHeader("Authorization: Basic ZjVjZDgxOTE0ODhjNDhjOThhYTFiY2JjODAxYmIwYTg6YTdmYWFhMDZkYTFhNGQ5MWJjZTllOTExOTk3MzYyZGY");
-  //     },
-  //     success: (data) => {
-  //       if (!data) {
-  //         alert('failure')
-  //         return;
-  //       }
-  //       alert('success')
-  //       localStorage.setItem('RefreshToken', data.refresh_token);
-  //     },
-  //     failure: (data) => {
-  //       if (!data) {
-  //         alert('failure')
-  //         return;
-  //       }
-  //       alert('failed')
-  //       localStorage.setItem('RefreshToken', data.refresh_token);
-  //     },
-  //   }).fail(alert('hi hi hi'));
+//   // Make a call using the token
+//   $.ajax({
+//     url: `https://accounts.spotify.com/api/token?grant_type=authorization_code&code=${code}&redirect_uri=http%3A%2F%2Flocalhost%3A3000%2Fregister`,
+//     type: "POST",
+//     beforeSend: (xhr) => {
+//       xhr.setRequestHeader("Authorization: Basic ZjVjZDgxOTE0ODhjNDhjOThhYTFiY2JjODAxYmIwYTg6YTdmYWFhMDZkYTFhNGQ5MWJjZTllOTExOTk3MzYyZGY");
+//     },
+//     success: (data) => {
+//       if (!data) {
+//         alert('failure')
+//         return;
+//       }
+//       alert('success')
+//       localStorage.setItem('RefreshToken', data.refresh_token);
+//     },
+//     failure: (data) => {
+//       if (!data) {
+//         alert('failure')
+//         return;
+//       }
+//       alert('failed')
+//       localStorage.setItem('RefreshToken', data.refresh_token);
+//     },
+//   }).fail(alert('hi hi hi'));
 
-  // }
+// }
 
-  // function getParameterByName(name, url) {
-  //   if (!url) url = window.location.href;
-  //   name = name.replace(/[\[\]]/g, "\\$&");
-  //   var regex = new RegExp("[?&]" + name + "(=([^&#]*)|&|#|$)"),
-  //     results = regex.exec(url);
-  //   if (!results) return null;
-  //   if (!results[2]) return "";
-  //   return decodeURIComponent(results[2].replace(/\+/g, " "));
-  // }
+// function getParameterByName(name, url) {
+//   if (!url) url = window.location.href;
+//   name = name.replace(/[\[\]]/g, "\\$&");
+//   var regex = new RegExp("[?&]" + name + "(=([^&#]*)|&|#|$)"),
+//     results = regex.exec(url);
+//   if (!results) return null;
+//   if (!results[2]) return "";
+//   return decodeURIComponent(results[2].replace(/\+/g, " "));
+// }
